// src/libs/enums/processBarEnums.ts
export enum StepProcessBar {
  FIRST = -1,
  POLICY_DETAILS = 0,
  SELECT_PLAN = 1,
  SELECT_ADD_ON = 2,
  PERSONAL_DETAIL = 3,
  COMPLETE_PURCHASE = 4,
}

export enum UserStep {
  POLICY_DETAILS = 'policy_details',
  SELECT_PLAN = 'select_plan',
  SELECT_ADD_ON = 'select_add_on',
  PERSONAL_DETAIL = 'personal_detail',
  COMPLETE_PURCHASE = 'complete_purchase',
  COMPLETE = 'complete',
}